Summarize basket

+++ src/main/java/com/joannahulek/checkout/component/CountableProduct.java
@@ -6,7 +6,8 @@ import java.math.BigDecimal;
 
 public class CountableProduct extends Product implements Countable{
   private int amount;
-    public CountableProduct(String name, double price, int amount) {
+
+    public CountableProduct(String name, BigDecimal price, int amount) {
         super(name, price);
         this.amount=amount;
     }
