+++ src/main/java/com/joannahulek/checkout/component/Product.java
@@ -4,9 +4,9 @@ import java.math.BigDecimal;
 
 public class Product {
     private String name;
-    private double price;
+    private BigDecimal price;
 
-    public Product(String name, double price) {
+    public Product(String name, BigDecimal price) {
         this.name = name;
         this.price = price;
     }
@@ -15,7 +15,7 @@ public class Product {
         return name;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 }
